create import_arbor method to avoid global scope import
