+++ bluepyopt/__init__.py
@@ -19,6 +19,25 @@ Copyright (c) 2016-2022, EPFL/Blue Brain Project
  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 """
 
+
+def import_arbor():
+    """Import arbor"""
+    try:
+        import arbor
+    except ImportError as e:
+        class arbor:
+            def __getattribute__(self, _):
+                raise ImportError("Exporting cell models to ACC/JSON, loading"
+                                  " them or optimizing them with the Arbor "
+                                  "simulator requires missing dependency"
+                                  " arbor. To install BluePyOpt with arbor,"
+                                  " run 'pip install bluepyopt[arbor]'.")
+    return arbor
+
+
+_arbor = import_arbor()
+
+
 # pylint: disable=W0611
 
 from ._version import get_versions
